requirements.txt
kivy==2.1.0
pillow>=9.0.0
